empl perform views
