+++ custom_reports/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # from . import filename_python_file_within_folder_or_subfolder
-from . import models
+from . import custom_report
+from . import employee_performance
 from . import ir_action
 from . import ir_ui_view
